Use more general apps for defaults
Using `ristretto` as image viewer since `swayimg` does not work on X11
and `nsxiv` does not work on wayland.

Also decided that since I'm using .desktop files just as fallback (They
will be used when I'm not in control), it's ok if the program is not
present. `ristretto` and `thunar` are EXTRA programs.